+++ core/src/raw/oio/mod.rs
@@ -35,14 +35,16 @@ pub mod into_reader;
 pub mod into_blocking_reader;
 
 mod write;
-pub use write::Append;
-pub use write::AppendOperation;
-pub use write::Appender;
 pub use write::BlockingWrite;
 pub use write::BlockingWriter;
 pub use write::Write;
 pub use write::WriteOperation;
 pub use write::Writer;
+
+mod append;
+pub use append::Append;
+pub use append::AppendOperation;
+pub use append::Appender;
 
 mod cursor;
 pub use cursor::Cursor;
