+++ core/src/services/sftp/error.rs
@@ -16,9 +16,11 @@
 // under the License.
 
 use openssh::Error as SshError;
-use openssh_sftp_client::{error::SftpErrorKind, Error as SftpClientError};
+use openssh_sftp_client::error::SftpErrorKind;
+use openssh_sftp_client::Error as SftpClientError;
 
-use crate::{Error, ErrorKind};
+use crate::Error;
+use crate::ErrorKind;
 
 #[derive(Debug)]
 pub enum SftpError {
